fix: apply suggestions from cr
