+++ docs/finance.md
@@ -292,7 +292,7 @@ TokenInsight 官方亦有提供 RSS，可参考 <https://api.tokeninsight.com/re
 
 ### 文章列表
 
-<Route author="moke8" example="/qianzhan-analyst/column/all" path="/qianzhan-analyst/column/:type" :paramsDesc="['分类，见下表']">
+<Route author="moke8" example="/qianzhan/analyst/column/all" path="/qianzhan/analyst/column/:type?" :paramsDesc="['分类，见下表']">
 
 | 全部  | 研究员专栏 | 规划师专栏 | 观察家专栏 |
 | --- | ----- | ----- | ----- |
@@ -302,7 +302,7 @@ TokenInsight 官方亦有提供 RSS，可参考 <https://api.tokeninsight.com/re
 
 ### 排行榜
 
-<Route author="moke8" example="/qianzhan-analyst/rank/week" path="/qianzhan-analyst/rank/:type" :paramsDesc="['分类，见下表']">
+<Route author="moke8" example="/qianzhan/analyst/rank/week" path="/qianzhan/analyst/rank/:type?" :paramsDesc="['分类，见下表']">
 
 | 周排行  | 月排行   |
 | ---- | ----- |
