--- lib/v2/qianzhan-analyst/router.js
@@ -1,4 +1,0 @@
-module.exports = (router) => {
-    router.get('/rank/:type?', require('./rank'));
-    router.get('/column/:type?', require('./column'));
-};
