+++ lib/v2/qianzhan/maintainer.js
@@ -1,0 +1,4 @@
+module.exports = {
+    '/analyst/column/:type?': ['moke8'],
+    '/analyst/rank/:type?': ['moke8'],
+};
