+++ lib/v2/qianzhan/radar.js
@@ -9,11 +9,17 @@ module.exports = {
                 target: (params) => {
                     if (params.html) {
                         const type = params.html.match(/\d+/)[0];
-                        return '/qianzhan-analyst/column/' + type;
+                        return '/qianzhan/analyst/column/' + type;
                     } else {
-                        return '/qianzhan-analyst/column/all';
+                        return '/qianzhan/analyst/column/all';
                     }
                 },
+            },
+            {
+                title: '排行榜',
+                docs: 'https://docs.rsshub.app/finance.html#qian-zhan-wang',
+                source: ['/analyst', '/'],
+                target: '/qianzhan/analyst/rank',
             },
         ],
     },
