+++ lib/v2/qianzhan/router.js
@@ -1,0 +1,4 @@
+module.exports = (router) => {
+    router.get('/analyst/column/:type?', require('./column'));
+    router.get('/analyst/rank/:type?', require('./rank'));
+};
